fix scale-types to types rename
