+++ projects/ngx-d3timeline/src/lib/axis/axis.service.ts
@@ -7,7 +7,7 @@ import { Line } from './line';
 import { Store } from '../store/store';
 import { OptionsService } from '../options.service';
 import { TimelineView } from '../view/timeline-view';
-import { Scale } from '../scale-types';
+import { Scale } from '../types';
 import { ResourceAxisTickRenderer } from './resources-axis/resource-axis-tick-renderer';
 import { TimeAxisTickRenderer } from './time-axis/time-axis-tick-renderer';
 import { TickRenderer } from './tick-renderer';
