+++ projects/ngx-d3timeline/src/lib/axis/time-axis/time-axis-tick-renderer.ts
@@ -1,4 +1,4 @@
-import { TimeScale } from '../../scale-types';
+import { TimeScale } from '../../types';
 import { TickRenderer } from '../tick-renderer';
 
 export class TimeAxisTickRenderer implements TickRenderer<TimeScale> {
